Valido si la agenda ya existe en sips
