+++ modules/turnos/controller/agenda.ts
@@ -1,10 +1,6 @@
-import * as moment from 'moment';
 import * as agendaModel from '../../turnos/schemas/agenda';
-import * as configPrivate from '../../../config.private';
-import * as sql from 'mssql';
-const MongoClient = require('mongodb').MongoClient;
+import * as moment from 'moment';
 
-let async = require('async');
 // Turno
 export function darAsistencia(req, data, tid = null) {
     let turno = getTurno(req, data, tid);
@@ -456,199 +452,3 @@ export function actualizarAgendas() {
     });
 
 }
-
-export function getAgendaSips() {
-    const url = configPrivate.hosts.mongoDB_main.host;
-    const coleccion = 'agendaSipsCache';
-
-    MongoClient.connect(url, function (err: any, dbMongo: any) {
-
-        if (err) {
-            console.log('Error conectando a mongoClient', err);
-            dbMongo.close();
-        }
-
-        dbMongo.collection(coleccion).find().toArray(function (err3, items) {
-
-            async.forEach(items, function (agenda, done) {
-
-                let offset = 0;
-                setTimeout(function () {
-                    console.log('Cursooo: ', agenda);
-                    grabaSips(agenda);
-                }, 5000 + offset);
-                offset += 10000;
-            });
-
-        });
-
-        dbMongo.close();
-    });
-}
-
-
-export function grabaSips(agendaSips: any) {
-
-    let objectId = agendaSips._id;
-    console.log("Objectttt " , objectId);
-    let estado = getEstadoAgendaSips(agendaSips.estado);
-    let codigoSisa = agendaSips.organizacion.codigo.sisa;
-    let dniProfesional = agendaSips.profesionales[0].documento;
-    let fecha = moment(agendaSips.horaInicio).format('YYYYMMDD');
-    let horaInicio = moment(agendaSips.horaInicio).utc().format('HH:mm');
-    let horaFin = moment(agendaSips.horaFin).utc().format('HH:mm');
-    let duracionTurno = agendaSips.bloques[0].duracionTurno;
-
-    let tipoPrestacion = agendaSips.tipoPrestaciones[0].conceptId;
-
-    let maximoSobreTurnos = 0;
-    let porcentajeTurnosDia = 0;
-    let porcentajeTurnosAnticipados = 0;
-    let citarPorBloques = 0;
-    let cantidadInterconsulta = 0;
-    let turnosDisponibles = 1;
-    let idMotivoInactivacion = 0;
-    let multiprofesional = 0;
-
-    agendaSips.idConsultorio = 273;
-
-    let datosSips = getDatosSips(codigoSisa, dniProfesional, tipoPrestacion);
-
-    let query;
-
-    Promise.all([datosSips]).then(values => {
-
-        let idEfector = values[0][0][0].idEfector;
-        let idProfesional = values[0][1][0].idProfesional;
-        let idEspecialidad = values[0][2][0].idEspecialidad;
-        let idServicio = values[0][2][0].idServicio;
-        let idTipoPrestacion = 0;
-
-        query = "insert into Con_Agenda (idAgendaEstado, idEfector, idServicio, idProfesional, idTipoPrestacion, idEspecialidad, idConsultorio, fecha, duracion, horaInicio, horaFin, maximoSobreTurnos, porcentajeTurnosDia, porcentajeTurnosAnticipados, citarPorBloques, cantidadInterconsulta, turnosDisponibles, idMotivoInactivacion, multiprofesional, objectId) values (" + estado + ", " + idEfector + ", " + idServicio + ", " + idProfesional + ", " + idTipoPrestacion + ", " + idEspecialidad + ", " + agendaSips.idConsultorio + ", '" + fecha + "', " + duracionTurno + ", '" + horaInicio + "', '" + horaFin + "', " + maximoSobreTurnos + ", " + porcentajeTurnosDia + ", " + porcentajeTurnosAnticipados + ", " + citarPorBloques + " , " + cantidadInterconsulta + ", " + turnosDisponibles + ", " + idMotivoInactivacion + ", " + multiprofesional + ", '" + objectId + "')";
-        console.log("Queryyyy ", query);
-
-        let query1 = "INSERT INTO dbo.CON_Turno ( idAgenda,  idTurnoEstado, idUsuario , idPaciente , fecha , hora , sobreturno , idTipoTurno , idObraSocial , idTurnoAcompaniante) VALUES  ( 75822 , 1 , 9739 ,  410551 , GETDATE() , '05:00' , 0 ,0 , 1 ,0 )";
-        
-        let connection = {
-            user: configPrivate.conSql.auth.user,
-            password: configPrivate.conSql.auth.password,
-            server: configPrivate.conSql.serverSql.server,
-            database: configPrivate.conSql.serverSql.database
-        };
-
-        (async function () {
-            try {
-                let pool = await sql.connect(connection);
-
-                const transaction = new sql.Transaction(pool);
-                transaction.begin(err => {
-                    // ... error checks 
-
-                    const request = new sql.Request(transaction)
-                    request.query(query, (err, result) => {
-                        // ... error checks
-                        if (err)
-                            return console.log("Errooo.", err.message);;
-
-
-                        transaction.commit(err1 => {
-                            // ... error checks 
-                            if (err1)
-                                return console.log('Transaction uncommitted.', err1);
-
-                            console.log('Transaction Committed.');
-                        });
-                    });
-
-                    // request.query(query1, (err, result) => {
-                    //     // ... error checks
-                    //     if (err)
-                    //         return console.log("Errooo.", err.message);;
-
-
-                    //     transaction.commit(err1 => {
-                    //         // ... error checks 
-                    //         if (err1)
-                    //             return console.log('Transaction uncommitted.', err1);
-
-                    //         console.log('Transaction Committed.');
-                    //     });
-                    // });
-
-                });
-
-
-            } catch (err) {
-                // ... error checks 
-            }
-        })();
-
-        sql.on('error', err => {
-            // ... error handler 
-        });
-    }, reason => {
-        console.log(reason)
-    });
-
-    //query = "insert into Con_Agenda (idAgendaEstado, idEfector, idServicio, idProfesional, idTipoPrestacion, idEspecialidad, idConsultorio, fecha, duracion, horaInicio, horaFin, maximoSobreTurnos, porcentajeTurnosDia, porcentajeTurnosAnticipados, citarPorBloques, cantidadInterconsulta, turnosDisponibles, idMotivoInactivacion, multiprofesional) values (" + agendaSips.agenda.estado + ", " + idEfector + ", " + agendaSips.agenda.idServicio + ", " + agendaSips.agenda.profesionales.idProfesional + ", " + agendaSips.agenda.idTipoPrestacion + ", " + agendaSips.agenda.tipoPrestaciones.idEspecialidad + ", " + agendaSips.agenda.idConsultorio + ", " + agendaSips.agenda.fecha + ", " + agendaSips.agenda.duracion + ", '" + agendaSips.agenda.horaInicio + "', '" + agendaSips.agenda.horaFin + "', " + agendaSips.agenda.maximoSobreTurnos + ", " + agendaSips.agenda.porcentajeTurnosDia + ", " + agendaSips.agenda.porcentajeTurnosAnticipados + ", " + agendaSips.agenda.citarPorBloques + " , " + agendaSips.agenda.cantidadInterconsulta + ", " + agendaSips.agenda.turnosDisponibles + ", " + agendaSips.agenda.idMotivoInactivacion + ", " + agendaSips.agenda.multiprofesional + ")";
-
-    // query = "INSERT INTO CON_Agenda (idAgendaEstado, idEfector, idServicio, idProfesional, idTipoPrestacion, idEspecialidad, idConsultorio, fecha, duracion, horaInicio, horaFin, maximoSobreTurnos, porcentajeTurnosDia, porcentajeTurnosAnticipados, citarPorBloques, cantidadInterconsulta, turnosDisponibles, multiprofesional ) VALUES  (" + agendaSips.agenda.estado + "," + agendaSips.turno.organizacion.codigo.sisa + "," + agendaSips.agenda.tipoPrestaciones.conceptId + ", " + agendaSips.agenda.profesionales.idProfesional + "," + agendaSips.agenda.idTipoPrestacion + "," + agendaSips.agenda.tipoPrestaciones.idEspecialidad + "," + agendaSips.agenda.idConsultorio + ", " + agendaSips.agenda.fecha + ", " + agendaSips.agenda.horaInicio + ", " + agendaSips.agenda.horaFin + ", " + agendaSips.agenda.maximoSobreTurnos + ", " + agendaSips.agenda.porcentajeTurnosDia + ", " + agendaSips.agenda.porcentajeTurnosAnticipados + ", " + agendaSips.agenda.citarPorBloques + ", " + agendaSips.agenda.cantidadInterconsulta + ", " + agendaSips.agenda.turnosDisponibles + ", " + agendaSips.agenda.idMotivoInactivacion + ", " + agendaSips.agenda.multiprofesional +  ")";
-
-}
-
-function getDatosSips(codigoSisa, dniProfesional, conceptId) {
-
-    var connection = {
-        user: configPrivate.conSql.auth.user,
-        password: configPrivate.conSql.auth.password,
-        server: configPrivate.conSql.serverSql.server,
-        database: configPrivate.conSql.serverSql.database
-    };
-
-    return new Promise((resolve: any, reject: any) => {
-        (async function () {
-            try {
-                let pool = await sql.connect(connection)
-                let result: any[] = [];
-
-                result[0] = await pool.request()
-                    .input('codigoSisa', sql.BigInt, codigoSisa)
-                    .query('select idEfector from dbo.Sys_Efector WHERE codigoSisa = @codigoSisa');
-
-                // console.dir(result1);
-
-                result[1] = await pool.request()
-                    .input('dniProfesional', sql.Int, dniProfesional)
-                    .query('SELECT idProfesional FROM dbo.Sys_Profesional WHERE numeroDocumento = @dniProfesional and activo = 1');
-
-                // console.dir(result2);
-
-                result[2] = await pool.request()
-                    .input('conceptId', sql.BigInt, conceptId)
-                    .query('SELECT E.idEspecialidad, S.idServicio FROM dbo.Sys_Especialidad E INNER JOIN dbo.Sys_Servicio S ON s.unidadOperativa = e.unidadOperativa WHERE E.conceptId_snomed = 268565007 AND s.activo = 1 ');
-
-                console.log("Especialidaddd ", result[2]);
-
-                resolve(result);
-            } catch (err) {
-                // ... error checks 
-                console.log(err);
-            }
-        })();
-    });
-}
-
-function getEstadoAgendaSips(estadoCitas) {
-    let estado: any;
-
-    if (estadoCitas === 'disponible' || estadoCitas === 'publicada') {
-        estado = 1;
-    } else if (estadoCitas === 'suspendida') {
-        estado = 2;
-    } else if (estadoCitas === 'codificada') {
-        estado = 4;
-    }
-
-    return estado;
-}
-
